Split utils tests into specific modules
